server/requirements.txt
fastapi
python-socketio
uvicorn
# setuptools